requirements.txt
flask
pywebview
selenium
pillow
numpy
webdriver-manager
